+++ command_attr/src/util.rs
@@ -2,11 +2,16 @@ use proc_macro2::Span;
 use proc_macro2::TokenStream as TokenStream2;
 use quote::{quote, ToTokens};
 use syn::{
+    parse_quote,
+    Type,
     braced, bracketed,
-    parse::{Parse, ParseStream, Result},
+    spanned::Spanned,
+    parse::{Parse, ParseStream, ParseBuffer, Result, Error},
     token::{Brace, Bracket, Mut},
-    Ident, Lit, Token, Type,
+    Ident, Lit, Token, ReturnType
 };
+use crate::structures::CommandFun;
+use crate::crate_name;
 
 pub trait LitExt {
     fn to_str(&self) -> String;
@@ -67,7 +72,7 @@ pub trait ParseStreamExt {
     fn try_parse<T: Parse>(&self) -> Result<T>;
 }
 
-impl<'a> ParseStreamExt for ParseStream<'a> {
+impl<'a> ParseStreamExt for ParseBuffer<'a> {
     fn try_parse<T: Parse>(&self) -> Result<T> {
         let stream = self.fork();
         let res = T::parse(&stream);
@@ -144,6 +149,13 @@ impl<T: Parse> Parse for RefOrInstance<T> {
     }
 }
 
+impl<T: Default> Default for RefOrInstance<T> {
+    #[inline]
+    fn default() -> Self {
+        RefOrInstance::Instance(T::default())
+    }
+}
+
 #[derive(Debug)]
 pub struct IdentAccess(pub Ident, pub Option<Ident>);
 
@@ -213,4 +225,291 @@ impl Parse for Array {
 
         Ok(Array(fields.into_iter().collect()))
     }
+}
+
+// FIXME: God give me the strength I need to shorten this.
+pub fn validate_declaration(fun: &mut CommandFun, is_help: bool) -> Result<()> {
+    if is_help && fun.args.len() > 6 {
+        return Err(Error::new(
+            fun.args.last().unwrap().span(),
+            "function's arity exceeds more than 6 arguments",
+        ));
+    } else if !is_help && fun.args.len() > 3 {
+        return Err(Error::new(
+            fun.args.last().unwrap().span(),
+            "function's arity exceeds more than 3 arguments",
+        ));
+    }
+
+    let context: Type = parse_quote!(&mut Context);
+    let message: Type = parse_quote!(&Message);
+    let args: Type = parse_quote!(Args);
+    let options: Type = parse_quote!(&'static HelpOptions);
+    let groups: Type = parse_quote!(&[&'static CommandGroup]);
+    let owners: Type = parse_quote!(HashSet<UserId>);
+
+    let cname = crate_name();
+    let context_path: Type = parse_quote!(&mut #cname::prelude::Context);
+    let message_path: Type = parse_quote!(&#cname::model::channel::Message);
+    let args_path: Type = parse_quote!(#cname::framework::standard::Args);
+    let options_path: Type = parse_quote!(&'static #cname::framework::standard::HelpOptions);
+    let groups_path: Type = parse_quote!(&[&'static #cname::framework::standard::CommandGroup]);
+    let owners_path: Type = parse_quote!(std::collections::HashSet<#cname::model::id::UserId>);
+
+    let ctx_error = "first argument's type should be `&mut Context`";
+    let msg_error = "second argument's type should be `&Message`";
+    let args_error = "third argument's type should be `Args`";
+    let options_error = "fourth argument's type should be `&'static HelpOptions`";
+    let groups_error = "fifth argument's type should be `&[&'static CommandGroup]`";
+    let owners_error = "sixth argument's type should be `HashSet<UserId>`";
+
+    match fun.args.len() {
+        0 => {
+            fun.args.push(Argument {
+                mutable: None,
+                name: Ident::new("_ctx", Span::call_site()),
+                kind: context_path,
+            });
+            fun.args.push(Argument {
+                mutable: None,
+                name: Ident::new("_msg", Span::call_site()),
+                kind: message_path,
+            });
+            fun.args.push(Argument {
+                mutable: Some(parse_quote!(mut)),
+                name: Ident::new("_args", Span::call_site()),
+                kind: args_path,
+            });
+
+            if is_help {
+                fun.args.push(Argument {
+                    mutable: None,
+                    name: Ident::new("_options", Span::call_site()),
+                    kind: options_path,
+                });
+
+                fun.args.push(Argument {
+                    mutable: None,
+                    name: Ident::new("_groups", Span::call_site()),
+                    kind: groups_path,
+                });
+
+                fun.args.push(Argument {
+                    mutable: None,
+                    name: Ident::new("_owners", Span::call_site()),
+                    kind: owners_path,
+                });
+            }
+        }
+        1 => {
+            if fun.args[0].kind != context {
+                return Err(Error::new(fun.args[0].span(), ctx_error));
+            }
+
+            fun.args.push(Argument {
+                mutable: None,
+                name: Ident::new("_msg", Span::call_site()),
+                kind: message_path,
+            });
+            fun.args.push(Argument {
+                mutable: Some(parse_quote!(mut)),
+                name: Ident::new("_args", Span::call_site()),
+                kind: args_path,
+            });
+
+            if is_help {
+                fun.args.push(Argument {
+                    mutable: None,
+                    name: Ident::new("_options", Span::call_site()),
+                    kind: options_path,
+                });
+
+                fun.args.push(Argument {
+                    mutable: None,
+                    name: Ident::new("_groups", Span::call_site()),
+                    kind: groups_path,
+                });
+
+                fun.args.push(Argument {
+                    mutable: None,
+                    name: Ident::new("_owners", Span::call_site()),
+                    kind: owners_path,
+                });
+            }
+        }
+        2 => {
+            if fun.args[0].kind != context {
+                return Err(Error::new(fun.args[0].span(), ctx_error));
+            }
+
+            if fun.args[1].kind != message {
+                return Err(Error::new(fun.args[1].span(), msg_error));
+            }
+
+            fun.args.push(Argument {
+                mutable: Some(parse_quote!(mut)),
+                name: Ident::new("_args", Span::call_site()),
+                kind: args_path,
+            });
+
+            if is_help {
+                fun.args.push(Argument {
+                    mutable: None,
+                    name: Ident::new("_options", Span::call_site()),
+                    kind: options_path,
+                });
+
+                fun.args.push(Argument {
+                    mutable: None,
+                    name: Ident::new("_groups", Span::call_site()),
+                    kind: groups_path,
+                });
+
+                fun.args.push(Argument {
+                    mutable: None,
+                    name: Ident::new("_owners", Span::call_site()),
+                    kind: owners_path,
+                });
+            }
+        }
+        3 => {
+            if fun.args[0].kind != context {
+                return Err(Error::new(fun.args[0].span(), ctx_error));
+            }
+
+            if fun.args[1].kind != message {
+                return Err(Error::new(fun.args[1].span(), msg_error));
+            }
+
+            if fun.args[2].kind != args {
+                return Err(Error::new(fun.args[2].span(), args_error));
+            }
+
+            if is_help {
+                fun.args.push(Argument {
+                    mutable: None,
+                    name: Ident::new("_options", Span::call_site()),
+                    kind: options_path,
+                });
+
+                fun.args.push(Argument {
+                    mutable: None,
+                    name: Ident::new("_groups", Span::call_site()),
+                    kind: groups_path,
+                });
+
+                fun.args.push(Argument {
+                    mutable: None,
+                    name: Ident::new("_owners", Span::call_site()),
+                    kind: owners_path,
+                });
+            }
+        }
+        4 => {
+            if fun.args[0].kind != context {
+                return Err(Error::new(fun.args[0].span(), ctx_error));
+            }
+
+            if fun.args[1].kind != message {
+                return Err(Error::new(fun.args[1].span(), msg_error));
+            }
+
+            if fun.args[2].kind != args {
+                return Err(Error::new(fun.args[2].span(), args_error));
+            }
+
+            if fun.args[3].kind != options {
+                return Err(Error::new(fun.args[3].span(), options_error));
+            }
+
+            fun.args.push(Argument {
+                mutable: None,
+                name: Ident::new("_groups", Span::call_site()),
+                kind: groups_path,
+            });
+
+            fun.args.push(Argument {
+                mutable: None,
+                name: Ident::new("_owners", Span::call_site()),
+                kind: owners_path,
+            });
+        }
+        5 => {
+            if fun.args[0].kind != context {
+                return Err(Error::new(fun.args[0].span(), ctx_error));
+            }
+
+            if fun.args[1].kind != message {
+                return Err(Error::new(fun.args[1].span(), msg_error));
+            }
+
+            if fun.args[2].kind != args {
+                return Err(Error::new(fun.args[2].span(), args_error));
+            }
+
+            if fun.args[3].kind != options {
+                return Err(Error::new(fun.args[3].span(), options_error));
+            }
+
+            if fun.args[4].kind != groups {
+                return Err(Error::new(fun.args[4].span(), groups_error));
+            }
+
+            fun.args.push(Argument {
+                mutable: None,
+                name: Ident::new("_owners", Span::call_site()),
+                kind: owners_path,
+            });
+        }
+        6 => {
+            if fun.args[0].kind != context {
+                return Err(Error::new(fun.args[0].span(), ctx_error));
+            }
+
+            if fun.args[1].kind != message {
+                return Err(Error::new(fun.args[1].span(), msg_error));
+            }
+
+            if fun.args[2].kind != args {
+                return Err(Error::new(fun.args[2].span(), args_error));
+            }
+
+            if fun.args[3].kind != options {
+                return Err(Error::new(fun.args[3].span(), options_error));
+            }
+
+            if fun.args[4].kind != groups {
+                return Err(Error::new(fun.args[4].span(), groups_error));
+            }
+
+            if fun.args[5].kind != owners {
+                return Err(Error::new(fun.args[5].span(), owners_error));
+            }
+        }
+        _ => unreachable!(),
+    }
+
+    Ok(())
+}
+
+pub fn validate_return_type(fun: &mut CommandFun) -> Result<()> {
+    let span = fun.ret.span();
+    let kind = match fun.ret {
+        ReturnType::Type(_, ref kind) => kind,
+        _ => unreachable!(),
+    };
+
+    let want: Type = parse_quote!(CommandResult);
+
+    if &**kind != &want {
+        return Err(Error::new(
+            span,
+            &format!(
+                "expected a result as a return type, but got `{}`",
+                quote!(#kind)
+            ),
+        ));
+    }
+
+    Ok(())
 }
