+++ src/client/dispatch.rs
@@ -199,7 +199,7 @@ fn dispatch_message<H>(
     });
 }
 
-#[allow(clippy::cyclomatic_complexity, unused_assignments, unused_mut)]
+#[allow(clippy::cognitive_complexity, unused_assignments, unused_mut)]
 fn handle_event<H: EventHandler + Send + Sync + 'static>(
     event: DispatchEvent,
     data: &Arc<RwLock<ShareMap>>,
